requirements.txt
# Core setup
python-dotenv
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
pydantic==2.11.5
langchain-groq==0.3.2
langchain-huggingface==0.2.0
